moving generateAudio to question controller and generateInterviewQUestionAudio method to question service
